requirements.txt
flask
mysql-connector-python
tesseract
